backend/requirements.txt
# CompressorAI v5 — Replit-Compatible Dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Auth
pyjwt==2.8.0
bcrypt==4.1.3

# Rate limiting
slowapi==0.1.9

# Env / config
python-dotenv==1.0.1

# Database
supabase==2.4.2

# Data processing
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2

# ML
scikit-learn==1.5.0
scipy==1.13.0

# Reports
reportlab==4.1.0

# Validation
pydantic[email]==2.7.0

# Scheduler
APScheduler==3.10.4
